+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.3.1
+## Changed
+- split RequestService::checkDocServiceUrl into checkDocServiceUrl, checkCommandService, and checkConvertService
+- add missing DocEditorConfig::setMode in DocEditorConfigService::getDocEditorConfig
+
 ## 1.3.0
 ## Added
 - ca-ES, cs-CZ, da-DK, hu-HU, id-ID, ro-RO, sq-AL, ur-PK empty file templates